Feature/auth

+++ app/Http/Controllers/UserController.php
@@ -2,17 +2,20 @@
 
 namespace App\Http\Controllers;
 
+use App\Notifications\DeleteUser;
+use App\Notifications\VerifyUser;
 use App\Rules\MatchesUserPassword;
-use App\Rules\Test;
-use App\User;
+use Carbon\Carbon;
+use Illuminate\Auth\AuthenticationException;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
     public function updatePassword(Request $request)
     {
-        $user = $request->user();
+        $user = Auth::user();
 
         $request->validate([
             'old_password' => ['required', new MatchesUserPassword($user)],
@@ -32,16 +35,39 @@ class UserController extends Controller
             'email' => ['required', 'string', 'email', 'max:255', 'unique:users'],
         ]);
 
-        $user = auth()->user();
+        $user = Auth::user();
+        $user->fill($data);
+        $user->generateVerificationToken();
+        $user->save();
 
-        $user->update($data);
+        $user->notify(new VerifyUser());
 
         return response()->noContent();
     }
 
-    public function destroy()
+    public function sendDelete()
     {
-        $user = auth()->user();
+        $user = Auth::user();
+
+        $user->generateDeleteToken();
+        $user->save();
+
+        $user->notify(new DeleteUser());
+
+        return response()->noContent();
+    }
+
+    public function destroy(Request $request)
+    {
+        $request->validate([
+            'delete_token' => 'required',
+        ]);
+
+        $user = Auth::user();
+
+        if (Carbon::now()->diffInMinutes($user->delete_token_generated_at) > 5 || $request->delete_token !== $user->delete_token) {
+            throw new AuthenticationException;
+        }
 
         $user->delete();
